+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  - Add kotlin support
  - Replace jsr305 with jetbrains annotations
  - Annotate interfaces null/not null for kotlin compatibility
+ - Remove retrolambda and target Java 8
 
 ### v0.0.16 - March 24th, 2019
  - Added new method to Mockspresso api `<T> void Mockspresso.inject(T, TypeToken<T>)`. This acts as a workaround when injecting a pre-existing object that is generic and has injected TypeVariables defined.